+++ src/ballistica/base/platform/apple/base_platform_apple.cc
@@ -4,7 +4,6 @@
 #include "ballistica/base/platform/apple/base_platform_apple.h"
 
 #if BA_XCODE_BUILD
-#include "ballistica/base/platform/apple/apple_utils.h"
 #include "ballistica/base/platform/apple/from_swift.h"
 #endif
 
@@ -62,6 +61,51 @@ void BasePlatformApple::DoOpenURL(const std::string& url) {
 #else
   // For non-xcode builds, go with the default (Python webbrowser module).
   BasePlatform::DoOpenURL(url);
+#endif  // BA_XCODE_BUILD
+}
+
+auto BasePlatformApple::OverlayWebBrowserIsSupported() -> bool {
+#if BA_XCODE_BUILD
+#if BA_OSTYPE_MACOS
+  return BallisticaKit::CocoaFromCpp::haveOverlayWebBrowser();
+#else
+  // TODO(ericf): Implement for uikit.
+  return BasePlatform::OverlayWebBrowserIsSupported();
+#endif  // BA_OSTYPE_MACOS
+
+#else
+  // Fall back to default for non-xcode apple builds.
+  return BasePlatform::OverlayWebBrowserIsSupported();
+#endif  // BA_XCODE_BUILD
+}
+
+void BasePlatformApple::DoOverlayWebBrowserOpenURL(const std::string& url) {
+#if BA_XCODE_BUILD
+#if BA_OSTYPE_MACOS
+  BallisticaKit::CocoaFromCpp::openURLInOverlayWebBrowser(url);
+#else
+  // TODO(ericf): Implement for uikit.
+  BasePlatform::DoOpenURLInOverlayBrowser(url);
+#endif  // BA_OSTYPE_MACOS
+
+#else
+  // For non-xcode builds, go with the default (Python webbrowser module).
+  BasePlatform::DoOverlayWebBrowserOpenURL(url);
+#endif  // BA_XCODE_BUILD
+}
+
+void BasePlatformApple::DoOverlayWebBrowserClose() {
+#if BA_XCODE_BUILD
+#if BA_OSTYPE_MACOS
+  BallisticaKit::CocoaFromCpp::closeOverlayWebBrowser();
+#else
+  // TODO(ericf): Implement for uikit.
+  BasePlatform::OverlayWebBrowserIsSupported();
+#endif  // BA_OSTYPE_MACOS
+
+#else
+  // Fall back to default for non-xcode apple builds.
+  BasePlatform::OverlayWebBrowserIsSupported();
 #endif  // BA_XCODE_BUILD
 }