+++ src/ballistica/base/python/methods/python_methods_base_1.h
@@ -1,7 +1,7 @@
 // Released under the MIT License. See LICENSE for details.
 
-#ifndef BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_APP_H_
-#define BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_APP_H_
+#ifndef BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_BASE_1_H_
+#define BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_BASE_1_H_
 
 #include <vector>
 
@@ -10,11 +10,11 @@
 namespace ballistica::base {
 
 /// Python methods related to app functionality.
-class PythonMethodsApp {
+class PythonMethodsBase1 {
  public:
   static auto GetMethods() -> std::vector<PyMethodDef>;
 };
 
 }  // namespace ballistica::base
 
-#endif  // BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_APP_H_
+#endif  // BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_BASE_1_H_