+++ src/ballistica/base/python/methods/python_methods_base_2.h
@@ -1,7 +1,7 @@
 // Released under the MIT License. See LICENSE for details.
 
-#ifndef BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_GRAPHICS_H_
-#define BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_GRAPHICS_H_
+#ifndef BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_BASE_2_H_
+#define BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_BASE_2_H_
 
 #include <vector>
 
@@ -10,11 +10,11 @@
 namespace ballistica::base {
 
 /// Graphics related individual python methods for our module.
-class PythonMethodsGraphics {
+class PythonMethodsBase2 {
  public:
   static auto GetMethods() -> std::vector<PyMethodDef>;
 };
 
 }  // namespace ballistica::base
 
-#endif  // BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_GRAPHICS_H_
+#endif  // BALLISTICA_BASE_PYTHON_METHODS_PYTHON_METHODS_BASE_2_H_